warehouse-service/pyproject.toml
[tool.poetry]
name = "datapooler"
version = "0.1.0"
description = "Manages Warehouse Migration and Related Services"
authors = ["Sam Stiyer <sam.stiyer@gmail.com>"]
readme = "README.md"
packages = [{ include = "datapooler" }]

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.2"
pydantic = "^2.8.2"
fastapi = ">=0.111.1"
typer = "^0.12.3"
psycopg2 = "^2.9.10"                                    # Requires libpq-dev on local
sqlalchemy-redshift = "^0.8.14"
boto3 = "^1.37.4"
orjson = "^3.10.6"
smart-open = "^7.0.4"
pydantic-settings = "^2.8.1"
setuptools = "78.1.1"
redshift-connector = "^2.1.2"
more-itertools = "^10.3.0"
hypercorn = "^0.17.3"
celery = "^5.4.0"
ipython = ">=8.28.0"
redis = "^5.1.1"
xgboost = "2.1.3"
polars = "^1.24.0"
usaddress = "^0.5.11"
ipdb = "^0.13.13"
scikit-learn = "1.5.2"
pytest = "^8.3.5"
kafka-python-ng = "^2.2.3"
sqlalchemy-utils = "^0.41.2"
json-log-formatter = "^1.1.1"
httpx = "^0.28.1"
asyncpg = "^0.30.0"
sqlalchemy = { version = "^1.4", extras = ["asyncio"] }
h11 = "0.16.0"
pytest-asyncio = "^1.0.0"
pre-commit = "^4.2.0"
uvloop = "^0.21.0"
ddtrace = "^3.10.2"
python-dotenv = "^1.1.1"
protobuf = "^6.32.0"
slack-sdk = "^3.36.0"
tenacity = "^9.1.2"


[tool.poetry.group.flake8.dependencies]
black = ">=24.4.2"
pytest = ">=8.3.5"
isort = ">=6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = ">=3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry.scripts]
datapooler = "datapooler.cli:app"
